+++ SA-Mod-Manager/Configuration/SADX/GameSettings.cs
@@ -6,6 +6,7 @@ using System.IO;
 using SAModManager.Common;
 using System.Text.Json;
 using System.CodeDom.Compiler;
+using System;
 
 namespace SAModManager.Configuration.SADX
 {
@@ -696,25 +697,59 @@ namespace SAModManager.Configuration.SADX
 		/// <returns></returns>
 		public static GameSettings Deserialize(string path)
 		{
-			if (File.Exists(path))
+			try
 			{
-				string jsonContent = File.ReadAllText(path);
+				if (File.Exists(path))
+				{
+					string jsonContent = File.ReadAllText(path);
 
-				return JsonSerializer.Deserialize<GameSettings>(jsonContent);
+					return JsonSerializer.Deserialize<GameSettings>(jsonContent);
+				}
+				else
+				{
+                    return new();
+                }
 			}
-			else
-				return new();
+            catch (Exception ex)
+            {
+                new MessageWindow(Lang.GetString("MessageWindow.DefaultTitle.Error"), Lang.GetString("MessageWindow.Errors.ProfileLoad") + "\n\n" + ex.Message, MessageWindow.WindowType.IconMessage, MessageWindow.Icons.Error).ShowDialog();
+            }
+            
+			return new();     
 		}
 
 		/// <summary>
 		/// Serializes an SADX GameSettings JSON File.
 		/// </summary>
 		/// <param name="path"></param>
-		public void Serialize(string path)
+		public void Serialize(string path, string profileName)
 		{
-			string jsonContent = JsonSerializer.Serialize(this, new JsonSerializerOptions { WriteIndented = true });
 
-			File.WriteAllText(path, jsonContent);
+			try
+			{
+				if (Directory.Exists(App.CurrentGame.ProfilesDirectory))
+				{
+					string jsonContent = JsonSerializer.Serialize(this, new JsonSerializerOptions { WriteIndented = true });
+					File.WriteAllText(path, jsonContent);
+				}
+				else
+				{
+					App.CurrentGame.ProfilesDirectory = Path.Combine(App.ConfigFolder, App.CurrentGame.gameAbbreviation);
+					Directory.CreateDirectory(App.CurrentGame.ProfilesDirectory);
+					if (Directory.Exists(App.CurrentGame.ProfilesDirectory))
+					{
+						path = Path.Combine(App.CurrentGame.ProfilesDirectory, profileName);
+						string jsonContent = JsonSerializer.Serialize(this, new JsonSerializerOptions { WriteIndented = true });
+						File.WriteAllText(path, jsonContent);
+					}
+				}
+			}
+			catch
+			{
+
+			}
 		}
+
+		
 	}
 }
